+++ src/SFML/Network/SocketImpl.hpp
@@ -27,15 +27,94 @@
 ////////////////////////////////////////////////////////////
 // Headers
 ////////////////////////////////////////////////////////////
-#include <SFML/Config.hpp>
-
+#include <SFML/Network/Socket.hpp>
 
 #if defined(SFML_SYSTEM_WINDOWS)
 
-#include <SFML/Network/Win32/SocketImpl.hpp>
+#include <SFML/System/Win32/WindowsHeader.hpp>
+
+#include <winsock2.h>
+#include <ws2tcpip.h>
 
 #else
 
-#include <SFML/Network/Unix/SocketImpl.hpp>
+#include <arpa/inet.h>
+#include <netdb.h>
+#include <netinet/in.h>
+#include <netinet/tcp.h>
+#include <sys/socket.h>
+#include <sys/types.h>
+#include <unistd.h>
 
 #endif
+
+#include <cstdint>
+
+
+namespace sf::priv
+{
+////////////////////////////////////////////////////////////
+/// \brief Helper class implementing all the non-portable
+///        socket stuff
+///
+////////////////////////////////////////////////////////////
+class SocketImpl
+{
+public:
+    ////////////////////////////////////////////////////////////
+    // Types
+    ////////////////////////////////////////////////////////////
+#if defined(SFML_SYSTEM_WINDOWS)
+    using AddrLength = int;
+    using Size       = int;
+#else
+    using AddrLength = socklen_t;
+    using Size       = std::size_t;
+#endif
+
+    ////////////////////////////////////////////////////////////
+    /// \brief Create an internal sockaddr_in address
+    ///
+    /// \param address Target address
+    /// \param port    Target port
+    ///
+    /// \return sockaddr_in ready to be used by socket functions
+    ///
+    ////////////////////////////////////////////////////////////
+    static sockaddr_in createAddress(std::uint32_t address, unsigned short port);
+
+    ////////////////////////////////////////////////////////////
+    /// \brief Return the value of the invalid socket
+    ///
+    /// \return Special value of the invalid socket
+    ///
+    ////////////////////////////////////////////////////////////
+    static SocketHandle invalidSocket();
+
+    ////////////////////////////////////////////////////////////
+    /// \brief Close and destroy a socket
+    ///
+    /// \param sock Handle of the socket to close
+    ///
+    ////////////////////////////////////////////////////////////
+    static void close(SocketHandle sock);
+
+    ////////////////////////////////////////////////////////////
+    /// \brief Set a socket as blocking or non-blocking
+    ///
+    /// \param sock  Handle of the socket
+    /// \param block New blocking state of the socket
+    ///
+    ////////////////////////////////////////////////////////////
+    static void setBlocking(SocketHandle sock, bool block);
+
+    ////////////////////////////////////////////////////////////
+    /// Get the last socket error status
+    ///
+    /// \return Status corresponding to the last socket error
+    ///
+    ////////////////////////////////////////////////////////////
+    static Socket::Status getErrorStatus();
+};
+
+} // namespace sf::priv
