+++ src/SFML/Network/Unix/SocketImpl.cpp
@@ -25,7 +25,7 @@
 ////////////////////////////////////////////////////////////
 // Headers
 ////////////////////////////////////////////////////////////
-#include <SFML/Network/Unix/SocketImpl.hpp>
+#include <SFML/Network/SocketImpl.hpp>
 
 #include <SFML/System/Err.hpp>
