+++ src/SFML/Network/Win32/SocketImpl.cpp
@@ -25,7 +25,7 @@
 ////////////////////////////////////////////////////////////
 // Headers
 ////////////////////////////////////////////////////////////
-#include <SFML/Network/Win32/SocketImpl.hpp>
+#include <SFML/Network/SocketImpl.hpp>
 
 #include <cstring>
